@@ -5,29 +5,36 @@
 part of firebase_auth;
 
 class FirebaseAuthApi {
-  const FirebaseAuthApi({@required FirebaseAuthService firebaseAuthService, @required String userAgent})
-      : assert(firebaseAuthService != null),
-        // assert(userAgent != null),
-        _firebaseAuthService = firebaseAuthService,
-        _userAgent = userAgent;
+  FirebaseAuthApi({@required IdentitytoolkitApi gitkit})
+      : assert(gitkit != null),
+        _requester = gitkit.relyingparty {
+    final RelyingpartyResourceApi a = gitkit.relyingparty;
+  }
 
-  final FirebaseAuthService _firebaseAuthService;
-  final String _userAgent;
+  final RelyingpartyResourceApi _requester;
 
   /// Calls the signUpNewUser endpoint, which is responsible anonymously signing up a user or signing in a user
   /// anonymously.
-  Future<BaseAuthResponse> signUpNewUser(BaseAuthRequest request) {
-    return _firebaseAuthService.signUp(request);
+  Future<SignupNewUserResponse> signupNewUser(IdentitytoolkitRelyingpartySignupNewUserRequest request) {
+    return _requester.signupNewUser(request);
   }
 
   /// Calls the getAccountInfo endpoint, which returns account info for a given account.
-  Future<List<UserDataResponse>> getAccountInfo(String firebaseAccessToken) {
-    return _firebaseAuthService.lookup(firebaseAccessToken);
+  Future<GetAccountInfoResponse> getAccountInfo(IdentitytoolkitRelyingpartyGetAccountInfoRequest request) {
+    return _requester.getAccountInfo(request);
   }
 
   /// Calls the createAuthURI endpoint, which is responsible for creating the URI used by the IdP to authenticate the
   /// user.
-  Future<CreateAuthUriResponse> createAuthUri(CreateAuthUriRequest request) {
-    return _firebaseAuthService.createAuthUri(request);
+  Future<CreateAuthUriResponse> createAuthUri(IdentitytoolkitRelyingpartyCreateAuthUriRequest request) {
+    return _requester.createAuthUri(request);
+  }
+
+  Future<GetOobConfirmationCodeResponse> getOobConfirmationCode(Relyingparty request) {
+    return _requester.getOobConfirmationCode(request);
+  }
+
+  Future<EmailLinkSigninResponse> emailLinkSignin(IdentitytoolkitRelyingpartyEmailLinkSigninRequest request) {
+    return _requester.emailLinkSignin(request);
   }
 }
