@@ -4,17 +4,19 @@
 
 part of firebase_auth;
 
-class SecureTokenService {
-  const SecureTokenService({@required HttpService service})
-      : assert(service != null),
-        _service = service;
+const String _userAgent = 'dart-api-client securetoken/v1/token';
 
-  final HttpService _service;
+class SecureTokenService {
+  SecureTokenService(Client client)
+      : assert(client != null),
+        _requester = ApiRequester(client, 'https://securetoken.googleapis.com/', '', _userAgent);
+
+  final ApiRequester _requester;
 
   /// Refresh a Firebase ID token by issuing an HTTP POST request to the securetoken.googleapis.com endpoint.
-  Future<ExchangeRefreshTokenResponse> refreshToken(ExchangeRefreshTokenRequest request) async {
-    final dynamic data = await _service.post('v1/token', request.json);
-    print(data);
-    return ExchangeRefreshTokenResponse.fromJson(data);
+  Future<SecureTokenResponse> refreshToken(SecureTokenRequest request) async {
+    return _requester
+        .request('/v1/token', 'POST', body: jsonEncode(request.json))
+        .then((dynamic data) => SecureTokenResponse.fromJson(data));
   }
 }
